requirements.txt
Flask==3.0.0
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
